account-service/requirements.txt
grpcio==1.70.0
grpcio-tools==1.70.0
protobuf==5.29.3
PyJWT==2.1.0
dnspython==2.7.0
pymongo==4.11.1
kafka-python==2.0.3
gunicorn==23.0.0
waitress==3.0.2
flask==3.1.0